Enable secondary_user_on_secondary_display for CtsAttributionSourceTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled.
The main use case is Automotive's multi-display Whole Cabin experience
where passengers (modeled as visible background users) can interact
with the display in front of them concurrently with the driver
(modeled as the the current user) interacting with driver's display.

Fixes: 365183253
Test: atest CtsAttributionSourceTestCases
Test: atest --user-type secondary_user CtsAttributionSourceTestCases
Test: atest --user-type secondary_user_on_secondary_display CtsAttributionSourceTestCases
Flag: TEST_ONLY

Change-Id: I87ae0d13e7b648177b6b09c089a755120aef3b50