Fix EmbeddedPhotoPickerTest#testOnSessionError_sessionIsNull
The test was flaky because it was asserting that the selected URI list
immediately reflects the selected items whereas in the actual scenario,
sometimes there is a minor delay between selecting the items and the
onUriPermissionGranted client invocation.

Also the test was asserting that the session immediately gets null on
force stopping the photopicker package.

The fix is to not involve any checks related to the selected URIs as
they are irrelevant to this test and instead assert that the
'onSessionError' client method is invoked within a given time frame
using a countDownLatch.

Flag: TEST_ONLY
Test: atest EmbeddedPhotoPickerTest#testOnSessionError_sessionIsNull
Bug: 369738313
Change-Id: I2c9485d22abfad3b33d0b1170bd11a2f10025440