Add tests for connections for AttributionSource chains
This change adds tests which verify permission is denied if an app in the
AttributionSource chain does not have permissions to open the camera, and that
permission is granted otherwise.

Permission denial for chains can only apply to Camera2 - Camera1 is unable to
implement this because its API is frozen and it always uses the application
context.

Bug: 190657833
Bug: 369841571
Test: CtsSecurityTestCases:CameraPermissionTest
Flag: com.android.internal.camera.flags.check_full_attribution_source_chain
Change-Id: I026331afb53dbdbfc2c1909710a0b55d6d3fce29