Change testAndroidBaselineProfile2021Support
Two changes are included here:

1) The CDD requirement was changed from C-1-13 to H-1-1
2) Waive this test for older devices

The Android 14 (U) CDD requires meeting the requirements of the
Android Baseline Profile 2021 (possible to meet with devices that
launched with Android 8 (Oreo)).  However, one device launched with
Android 12 (S) does not meet those requirements, and fails the
testAndroidBaselineProfile2021Support test.  We will modify the test
to allow that device to pass the test.

Bug: 365661235
Flag: TEST_ONLY
Test: tools/cts-tradefed run commandAndExit cts -m CtsGraphicsTestCase
Change-Id: I7e5bc1152055469a9975865a676457176c600858