Enable secondary_user_on_secondary_display on CtsActivityRecognitionTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on devices
that have config_multiuserVisibleBackgroundUsers enabled. The main use
case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the display
in front of them concurrently with the driver (modeled as the the
current user) interacting with driver's display.

Bug: 364413990
Test: atest --user-type secondary_user_on_secondary_display CtsActivityRecognitionTestCases
Test: atest CtsActivityRecognitionTestCases
Flag: TEST_ONLY
Change-Id: Id7c36e16c7f6eac6fc7e395011285bf8a72020bf