CTS tests for TV_IMPLICIT_ENTER_PIP permission
Checks that there is no difference to the auto-enter behavior if the
permission is held and adds new tests to make sure PiP is not entered if
the permission is not held.
Includes testing for setAutoEnter, onPause, onBackPressed and
onUserLeaveHint and one test that explicit PiP entry still works.

Bug: 283115999
Test: atest CtsWindowManagerDeviceOther:PinnedStackTests
Flag: android.app.enable_tv_implicit_enter_pip_restriction

Change-Id: I7254a54a0553b2ccb5014d199cfd04223aa2ad70