Enable secondary_user_on_secondary_display for CtsAppSearchTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled.
The main use case is Automotive's multi-display Whole Cabin experience
where passengers (modeled as visible background users) can interact
with the display in front of them concurrently with the driver
(modeled as the the current user) interacting with driver's display.

Fixes: 373484460
Test: atest CtsAppSearchTestCases
Test: atest --user-type secondary_user  CtsAppSearchTestCases
Test: atest --user-type secondary_user_on_secondary_display CtsAppSearchTestCases
Flag: TEST_ONLY

Change-Id: I6594f5e844c3c95b61e38b2ff1dc32fe64020974