Skip testStartUserOnDisplayAndStopUser if no display is available.
On a single display device with isVisibleBackgroundUsersEnabled true,
the test failed due to no available display.
The test is about starting and stopping a user on a secondary display,
so it should be skipped if no secondary display is available.

Fixes: 374339299
Flag: TEST_ONLY
Test: atest android.car.cts.CarUserManagerTest
(Tested on both single-display and multi-display emulator)

Change-Id: I73e7b29febb547ea6d0a370fd317f98914fcde6c