Add CTS test for generated preview persistence
This change adds a CTS test for generated preview persistence. It tests
that a preview file is written on setWidgetPreview, and deleted on
removeWidgetPreview.

This change also updates testSetGeneratedPreview to wait for the
providersChanged callback before testing the result of setWidgetPreview
or removeWidgetPreview.

Bug: 364420494
Test: atest "CtsAppWidgetTestCases:AppWidgetTest#testGeneratedPreviewPersistence,testSetGeneratedPreview"
Flag: android.appwidget.flags.remote_views_proto

Change-Id: I4db2bc95bd15e553cbf5c5daf792cbf2d48ad660