Enable --secondary_user_on_secondary_display for CtsMediaAudioTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled. The
main use case is Automotive's multi-display Whole Cabin experience where
passengers (modeled as visible background users) can interact with the
display in front of them concurrently with the driver (modeled as the
the current user) interacting with driver's display.

Since microphone is not fully supported on secondary display users, so
related tests are disable for secondary user on secondary display.

Bug: 356711219
Flag: EXEMPT resource only update
Test: atest CtsMediaAudioTestCases
Test: atest --user-type secondary_user CtsMediaAudioTestCases
Test: atest --user-type secondary_user_on_secondary_display CtsMediaAudioTestCases
Change-Id: Ib91fc87abfb84f86cf818b7cf483e5668d5474e2