Enable secondary_user_on_secondary_display for CtsRsBlasTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled.
The main use case is Automotive's multi-display Whole Cabin experience
where passengers (modeled as visible background users) can interact
with the display in front of them concurrently with the driver
(modeled as the the current user) interacting with driver's display.

Fixes: 374403971
Test: atest CtsRsBlasTestCases
Test: atest --user-type secondary_user CtsRsBlasTestCases
Test: atest --user-type secondary_user_on_secondary_display CtsRsBlasTestCases
Flag: TEST_ONLY

Change-Id: Ie1ce804ff40d829294f6f86b845bf9564bbc4f58