Move common HVAC VehiclePropertyVerifiers to same source
This module is used by both CTS and ATS, so we have to handle different Android versions. For HVAC_TEMPERATURE_DISPLAY_UNITS the PERMISSION_READ_DISPLAY_UNITS was added starting Android U. For HVAC_FAN_DIRECTION, in Android U we throw an exception if an unwritable value tries to be set.

Flag: EXEMPT test-only
Bug: 366372483
Fix: 374493958
Test: atest CtsCarTestCases:CarPropertyManagerTest
Test: run ats-in-car -m AtsInCarInteractiveSensorDeviceTests
Change-Id: I009d96f98d43173e1e52d8d423c7172ce0c47205