Fix initial value issue for adas property.
Fixes an issue where the initial value is not correctly stored
and restored for adas property.

When the initial adas property enable state (e.g. CRUISE_CONTROL_ENABLED)
is false, the CRUISE_CONTROL_TYPE should be -2. We store the initial
enabled state as false, then we enable it, changing the TYPE to 2.
We then store the type as 2.

After the test case, we restore enable to false, and type to 2 which
is not a valid combination.

We should store the initial type value before we change the enable
state.

Flag: EXEMPT CTS
Test: atest android.car.cts.CarPropertyManagerTest
Bug: 375687678
Change-Id: I20ed05969b548a22adf5d36d2281a407cd617593