Log exceptions instead of throwing them.
Exceptions will no longer cause the collector to report an error.
This is in line with the majority of the other collectors.

Collector errors don't erase/invalidate the collector's (possibly
incomplete) JSON output, so this is no-op in terms of the collected
data. Thus, the only real effect of this change is that the
collector's logcat output is slightly more useful if there's an
exception.

This is a tradeoff since it's less obvious when running the collector
locally if something goes wrong. However, a local run should only
be deemed "successful" if the actual JSON output contains the
expected data and a failure will result in the expected fields not
being populated.

Bug: 371159583
Fix: 335906045
Test: ran on Pixel 7 Pro EVT, collector succeeds and the exception
      about failed ID attestation appears in logcat with the
      "AttestationDeviceInfo" tag

Change-Id: I74da9b98840bef2a17ec68472e2dea40a0a97d09