Add CTS tests for non-UI context in WM extension.
This commit adds CTS tests to ensure that the WindowManager extension
throws an IllegalArgumentException when the given context to
`WindowLayoutComponent#getCurrentWindowLayoutInfo` or
`WindowLayoutComponent#addWindowLayoutInfoListener` is not a UI context.

This helps to enforce the requirement that these methods must be called
with a valid UI context, ensuring that the `WindowInfoTracker` in
WindowManager Jetpack behaves as expected.

Bug: 337820752
Flag: com.android.window.flags.wlinfo_oncreate
Test: atest CtsWindowManagerJetpackTestCases:ExtensionWindowLayoutComponentTest
Change-Id: I907439e50dcae6e8e69f5f0b9caad4d4a2fd812c