Add CTS for setStreamWithCrops and setBitmapWithCrops
With the multi crop project, all older overloads of setStream and
setBitmap will be redirected to the new setBitmapWithCrops and
setStreamWithCrops methods. These methods thus deserve end-to-end tests
for the basic cases.

We test the method when one or two crops are provided. In these tests
the crop matches exactly the size of the display. We check that in such
a case the crop is not adjusted or modified, and that all methods
related to multi-crop are consistent.

Flag: com.android.window.flags.multi_crop
Flag: android.app.customization_packs_apis
Bug: 320972046
Test: atest WallpaperManagerTest
Change-Id: I5f94bcef70eb13690dc411c722e1334d21ee8846