Revert back to the initial user and remove test users in CarUserManagerHostTest
CarUserManagerHostTest switches users in its tests, but does not switch
back to the initial user that it started on. This is a problem because
the test APKs are installed on the initial user. As a result, tests that
depend on these APKs cannot load them when they run after
CarUserManagerHostTest. This CL adds tracking of the initial user and
existing users using a before method and updates the correct user state
using an after method.

Test: atest CtsCarHostTestCases
Flag: EXEMPT refactor
Fixes: 360978451
Change-Id: I5ceff11a839412cd5bed930be3b681cb53bee1f9