Enable secondary_user_on_secondary_display CtsCompanionDeviceManagerCoreTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled.
The main use case is Automotive's multi-display Whole Cabin experience
where passengers (modeled as visible background users) can interact
with the display in front of them concurrently with the driver
(modeled as the the current user) interacting with driver's display.

Fixes: 376559593
Test: atest CtsCompanionDeviceManagerCoreTestCases
Test: atest --user-type secondary_user CtsCompanionDeviceManagerCoreTestCases
Test: atest --user-type secondary_user_on_secondary_display CtsCompanionDeviceManagerCoreTestCases
Flag: TEST_ONLY

Change-Id: Ibbd93d2940ddd88cd0af97cddd41c3dc55bd4427