CtsWidgetTestCases:operate: A patch for MediaControllerTest to fix a failure in testOnBackInvokedCallback()
Regarding the peculiar phenomenon where the MediaControllerTest#testOnBackInvokedCallback case passes in the full package test but fails in the individual test, we provide the following reasons and explanations:
For the testOnBackInvokedCallback case, there exist ViewRootImpls, specifically VRI1 corresponding to the test case Activity and VRI2 corresponding to the MediaController.
When executing the case in full, the window focus sequence is entering Launcher -> VRI1 -> executing other cases -> executing the target case, meaning the focus can stably change from VRI1 to VRI2.  In the case of individual execution, the window focus sequence is Launcher -> VRI1 -> VRI2.
This means that when the case is executed individually, the window's entry into VRI1 might be delayed.  Since the focus switch in VRI is executed via post, when it comes to judging the loss of focus in VRI1, it might be during the "focus just left Launcher and has not yet entered VRI1" rather than "focus leaving VRI1 entering VRI2".   It leads to the misjudgment of the case "MediaController becomes visible and has focus" as passed!
Subsequently, the focus then enters VRI1, causing KEYCODE_BACK to be mistakenly sent to VRI1.  Therefore, after the focus enters VRI2, it can no longer receive the KEYCODE_BACK event, leading to the timeout failure of the case "KEYCODE_BACK can still dismiss the MediaController".
However, when executing the full case, by the time it gets to the target case, the focus is correctly on VRI1, thus ensuring the judgment that VRI1 has no focus can be equivalent to "the focus leaves VRI1 and enters VRI2".  Afterwards, the KEYCODE_BACK event can be correctly sent to VRI2, allowing the case to pass normally.
In summary, this can be considered a poor consideration of the conditions of judgment in the critical state when executing the case individually.
Therefore, a possible fix could be:
When testing "MediaController becomes visible and has focus", the patch judges the loss of focus in VRI1 while also judging the gain of focus in VRI2, to obtain a stable and expected focus state to pass the test normally (especially when the case is executed individually).

failure message:
junit.framework.AssertionFailedError: unexpected timeout
	at junit.framework.Assert.fail(Assert.java:50)
	at junit.framework.Assert.assertTrue(Assert.java:20)
	at com.android.compatibility.common.util.PollingCheck.run(PollingCheck.java:81)
	at com.android.compatibility.common.util.PollingCheck.waitFor(PollingCheck.java:146)
	at android.widget.cts.MediaControllerTest.testOnBackInvokedCallback(MediaControllerTest.java:196)

Bug: 374420388
Change-Id: I80816508f4490399df2eed17fa8f0675084c9e64
Test: run cts -m CtsWidgetTestCases -t android.widget.cts.MediaControllerTest#testOnBackInvokedCallback