Add notification when satellite availability changes
When satellite availaibility changes, we need to notify the state to users so that they're aware of it.
Specifically, when the satellite connectivity is not available, we need to show users the reason so that
they can act accordingly. For example, if we show that satellite connectivity is not available because
location settings is disabled, users can enable location settings and check availability again.

Test: Manual
Bug: 376694756
Flag: EXEMPT bugfix
Change-Id: I010573ef9fdb6d50b8e9b8780e5c95d658a4dbb7