Skip a test if config_hsumBootStrategy isn't 0
Test setBootUser_providedUserIsNotSwitchable_Hsum is not aware of the
new config_hsumBootStrategy, and assumes the boot user is always
previous foreground user.

When config_hsumBootStrategy is not 0, the boot user is either user 0 or
user 10, which breaks the above assumption.

This CL adds RequireResourcesIntegerValue to skip the test based on the
config_hsumBootStrategy value.

Bug: 378460323
Test: atest CtsMultiUserTestCases:android.multiuser.cts.UserManagerTest#setBootUser_providedUserIsNotSwitchable_Hsum
Test: atest com.android.bedstead.harrier.MainAnnotationExecutorTest#requireResourcesIntegerValueIsTrue_resourceValueIs1
Flag: EXEMPT test only

Change-Id: Ie7510a970d214a5f106a2c4b9a47bd068cdb2bf9