Add some tolerance to CTS crop tests
If the bitmap is way larger than the screen, we downsample it. Then when
trying to find the crops for the original bitmap, there can be some
rounding errors since to compute the original crops, we first compute
the crops for the downsampled bitmap (which are rounded) then scale
them back to the original bitmap size.

The tolerance was missing for one equality check.

Flag: com.android.window.flags.multi_crop
Fixes: 378744637
Test: presubmit
Test: atest WallpaperManagerTest
Change-Id: I01e5a9c73079390e6b9d51ef128cee01243cfa7b