Skip StylusButtonInputEventTest for background user
The StatusBarManagerService prevents background users from sending system keys to the status bar. This restriction is implemented
in the handleSystemKey method (specifically within enforceValidCallingUser). Consequently, tests that rely on sending system keys
from a background user will fail.

For example, this test will fail when executed in secondary_user_on_secondary_display mode. In this mode, the test runs as a
visible background user designated as the secondary user on a secondary display.

Bug: 335725205
Test: atest --user-type secondary_user_on_secondary_display CtsInputTestCases:android.input.cts.StylusButtonInputEventTest
Test: atest --user-type secondary_user CtsInputTestCases:android.input.cts.StylusButtonInputEventTest
Test: atest CtsInputTestCases:android.input.cts.StylusButtonInputEventTest
Flag: TEST_ONLY
Change-Id: I32c5f5e39fac29185773ed520233133c0d6fcc85