Fix BlurPixelVerifier to use the correct Gaussian blur equation.
Updates it to extend PerPixelBitmapVerifier to benefit from pixel
error logging and difference image dumping.

The previous implementation appeared to use a linear weighting between
the src and dst colors and relied on heavy tolerances to allow the
transitioning pixels to pass the verifier.

This approximation stopped passing when making changes to Skia-Ganesh's blur algorithm (removing SK_USE_LEGACY_BLUR_GANESH here
https://googleplex-android-review.git.corp.google.com/c/platform/external/skia/+/30376854). With
this change, both the new and old algorithms pass.

Additionally, updating to the new test caught a mistake in
testViewRenderNodeBlurEffect() where it was never adding anything to
trigger a blur. The old verifier was incorrectly identifying it as
passing anyways.

The new BlurPixelVerifier replaces the need for a RegionVerifier or
a separate ColorVerifier on the interior (the solid color interior's
bounds were also too large given differences in Android's definition
of radius->sigma and the actual influence of a given sigma; but it
was passing because the DEFAULT_THRESHOLD value is pretty tolerant).


Flag: TEST_ONLY
Bug: b/238761355
Change-Id: I57942d82c1e80c8f60467e79ea994384f5557130