Make WindowInputTests multi-user and multi-display aware (2/2)
This changelist ensures that OverlayTestService creates the overlay view
on the test user main display, so it can pass when running for the
secondary user (visible background user) assigned to a secondary
display.

EVENT_FLAGS_WAIT_TIME is also increased to allow better stability on slower surfaces.

Flag: TEST_ONLY
Test: atest CtsWindowManagerDeviceInput:android.server.wm.input.WindowInputTests
Test: atest --user-type secondary_user CtsWindowManagerDeviceInput:android.server.wm.input.WindowInputTests
Test: atest --user-type secondary_user_on_secondary_display CtsWindowManagerDeviceInput:android.server.wm.input.WindowInputTests
Bug: 378456865
Change-Id: Id293cc504481e0d89ba7c6eb7618c001f368a9e3