bt(snippet): Dont hold shell identity on initialization
We have a helper method to adopt and drop shell identity when needed in the context of handling an RPC call. Holding the shell identity globally breaks when the snippet is used in tests which uses multiple snippets since UiAutomation is a singleton and cannot be shared across multiple
processes.

Bug: 380329337
Change-Id: I85cf4327909e1b9d4e72e9578b4a45f7828065b0
Test: atest CtsBluetoothMultiDevicesTestCases MultiDeviceRangingTestCases