Change again how HiddenApiTest is ignored for JUnit 3
The change in
https://android-review.googlesource.com/c/platform/cts/+/3367539
didn't fix the issue. Even though the test function was made
private, the framework still tried to run the tests and failed the
access checks.

If I try changing the function names, then the test run fails
because there is no test to run. In the end, I've let the test
functions as they are, but remove their contents. This way the tests
pass.

Unfortunately I can't find another way to skip the tests, or
consistently log that the tests are skipped.

Bug: 378992956
Test: run cts --include-filter "CtsHiddenApiBlocklistApi27TestCases android.signature.cts.api.api27.HiddenApiTest"
Merged-In: Idd4df74d17cd396a1e751279e221497889702ed4

Change-Id: I2bf9962c7bc9bef6b115b876f740d6e9e45c16a4