Make WindowUntrustedTouchTest multi-user and multi-display aware
This changelist updates WindowUntrustedTouchTest to pass the display ID associated with the test activity (mActivity) to UntrustedTouchTestService. This allows UntrustedTouchTestService to create views with the correct window manager and window context.

Previously, this test failed when running for a secondary user on a secondary display (a visible background user on a secondary display). This failure occurred because DEFAULT_DISPLAY is not the main display for such users. Instead, their main display ID is assigned when the background visible user is created. This change ensures the test uses the correct display ID in all scenarios.

Flag: TEST_ONLY
Test: atest CtsWindowManagerDeviceInput:android.server.wm.input.WindowUntrustedTouchTest
Test: atest --user-type secondary_user CtsWindowManagerDeviceInput:android.server.wm.input.WindowUntrustedTouchTest
Test: atest --user-type secondary_user_on_secondary_display CtsWindowManagerDeviceInput:android.server.wm.input.WindowUntrustedTouchTest
Bug: 378456865
Change-Id: I16c381291939f9b38307faad0c3a58abc2760755