Add CTS test for Dependency Installer's multi-user support
We now maintain a user to Dependency Installer service map in the
system. So for each user, we bind to Dependency Installer service
installed for that user.

The test app itself is the dependency installer service in our cts
tests. But when the system binds to DIS in another user, the bound
service is not the instrumented process. We use this trait to figure out
if we are running the multi-user test or not.

Another problem: PackageManagerShellCommandInstallTest(PMSCIT)
communicates with TestDependencyInstallerService(TDIS) via
SharedPreferences. This only works when we bind to TDIS in current user
where TDIS is using the same internal storage as PMSCIT. Under a new
user, TDIS uses separate internal storage. So we are limited in how much
we can communicate across users.

Bug: 380402322
Bug: 372862145
Test: atest PackageManagerShellCommandInstallTest
FLAG: android.content.pm.sdk_dependency_installer
Change-Id: I53bb87d02b18c0fafdfdba9730d7d48da4ee493c