Disable DeviceConfig#getAdbWritableNamespaces test on ravenwood
Ravenwood does not yet support the DeviceConfig getWritableNamespaces
API, so when the new flag for this API is enabled, the ravenwood
test fails because a null result is returned. This commit disables
the test for this new API on ravenwood until the new API is
available in ravenwood.

Fixes: 380997267
Flag: TEST_ONLY
Test: atest CtsDeviceConfigTestCasesRavenwood:DeviceConfigApiTests
Change-Id: I0c8d2e929a80077879446a45c82f170787b75ebd