Add host side test for generated preview persistence
The device test for generated preview persistence does not work on
devices where shell cannot access the system appwidget folder. This
change removes that tests and adds a host side test for this instead.

This test sets a preview, reboots the device, and checks that the
preview is still available.

Bug: 378016529
Test: atest CtsAppWidgetHostTestCases
Flag: android.appwidget.flags.remote_views_proto
Change-Id: Icc37c1266ac0ae4c001596646212a2dd365e7bfc