Account for the audio permissions propagation
Workaround test fix until b/383048413 is fixed to use
PermissionUpdateBarrierRule.

An intentional delay was added while propagating the audio permissions, the tests using audio permission should account for it. App impact should be minor since the permissions are given at install and there is enough delay when an app starts for the audio permissions to be propagated, but for tests there is some flakyness, so allow for the right amount of time for the audio service to reach a stable state.

Bug: 375464080
Bug: 382632836
Test: atest CtsVirtualDevicesAudioTestCases:android.virtualdevice.cts.audio.VirtualAudioTest --iterations=100
Test: atest CtsVirtualDevicesAudioTestCases:android.virtualdevice.cts.audio.VirtualAud
ioPermissionTest#audioInjection_virtualDeviceWithManualAudioPolicy_permissionGrantedOnlyOnDefaultDevice --iterations=100
Flag: EXEMPT cts test fix

Change-Id: Iacc95082e390f040b90c8e4227170221615e7822