Ensure MockIme is tall enough to receive system window insets
The height of the content of MockIme can be shorter than the bottom
inset of system windows. It would require serveral iterations to make
the window of MockIme grow tall enough to receive the full system
window insets. The iteration:

1. MockIme receives the bottom inset.
2. MockIme sets the bottom inset to the bottom padding.
3. If MockIme window becomes taller, go to 1. Else, stop.

This CL make sure the process happens when the window is created, as
before edge-to-edge, not when MockIme is requested to show.

Fix: 382565099
Flag: com.android.window.flags.enforce_edge_to_edge
Test: atest WindowInsetsControllerTests
            #testDispatchApplyWindowInsetsCount_ime
Change-Id: I0672537dc13b1026745c97239cded9b785d98f5e