Fix ListViewTest stretch scrolling tests
Fixes: 380234259
Fixes: 378133975
Fixes: 377853670
Fixes: 378002689
Fixes: 379721387

ListView isn't necessarily ready right after it has been attached
to the hierarchy. The adapter needs to finish loading and the
ListView lays out the content. Tests should wait for the content
to load.

This waits for the first draw after the ListView has been made
visible. It also moves the ListView to the center of the screen
so that it can't bump into the system gesture area.

Flag: No flag needed to fix a flaky test

Test: ran tests locally several times
Change-Id: I137183fef9570193e387918445693591401f0ac9