Adjusting test case to not rely on receiving new insets, but being the same as before
With the flag turned on: When coming back from another app, the previous activity won't get new insets (as they're unchanged).
In this case, the test should check if the IME is visible and has the same insets as when going to the other app.

Original patch:
 From 1193ab826232af3a33f06104a71f3e7b2d1fdbcc Mon Sep 17 00:00:00 2001
From: Felix Stern <fstern@google.com>
Date: Tue, 01 Oct 2024 15:07:21 +0000
Subject: [PATCH] Adjusting test case to not rely on receiving new insets, but being the same as before

With the flag turned on: When coming back from another app, the previous activity won't get new insets (as they're unchanged).
In this case, the test should check if the IME is visible and has the same insets as when going to the other app.

Test: atest KeyboardVisibilityControlTest#testImeVisibilityWhenImeTransitionBetweenActivities
Flag: android.view.inputmethod.refactor_insets_controller
Bug: 298172246
Change-Id: Iecd50a7652b0618c1f454a094424349487ad79aa
---

Test: atest KeyboardVisibilityControlTest#testImeVisibilityWhenImeTransitionBetweenActivities
Flag: android.view.inputmethod.refactor_insets_controller
Bug: 378517010
Merged-In: Iecd50a7652b0618c1f454a094424349487ad79aa
Change-Id: Iecd50a7652b0618c1f454a094424349487ad79aa