SoundPoolTest: improve robustness to testLoadMore
Issue a retry: In the rare case where SoundPool is processing all streams,
we do not block but return 0 when play is called.  The app is free to retry
or ignore playing the sound, as the sounds are short and may be ignored
due to priority or other reasons.

The underlying goal of SoundPool is to have a "fire and forget" player
which plays sounds without blocking, and minimizes AudioTrack (stream)
usage.

Flag: EXEMPT bugfix
Test: atest SoundPoolAacTest SoundPoolHapticTest SoundPoolMidiTest SoundPoolOggTest
Bug: 380538319
Bug: 417711746
Change-Id: I89bb91abed582b0cb519de0cd0f067da7139f987