Skip the test cases on smaller display devices
Depending on the height/weight of the screen size, The Autofill UI may
not visible when the smaller display device is rotated to landscape.

Add a condition to check if the size is at least 480dp.
(This value is chosen to be the same as the minimum height size for
medium screen in dp)
If it's smaller than 480dp then should skip these test cases.

Bug: 380312360
Test: atest CtsAutoFillServiceTestCases
Change-Id: I5ecb01cb7ade03b8ae345b6dd7403014a859dc26