[ITS][JCA] use UI elements to change flash mode instead of static count
Details: Low light boost mode has been added as a flash mode for devices that support it, which means that static number of clicks may not result in the desired result for certain devices. Changed behavior to use UI elements to confirm that the correct flash mode has been set.

Change-Id: Ie0458b74bf1e7e5f5ea12503792f023623c532d7
Flag: TEST_ONLY
Bug: 383392277
Test: ran test_auto_flash