Fix broken test in AbsListViewTest
Fixes: 383571612

Flag: Test fixes, no flag needed.

Two different ListViews are in the same FrameLayout, and one of them
is being placed over the other. When the test clicks on an item,
it hits the wrong ListView.

This CL makes one of the ListViews GONE by default so that tests
don't accidentally touch it.

Test: ran broken test
Change-Id: I6cfb6002cdff31a0b7ccea1d17efceecb78950d4