Add shouldShowImeSwitcherButton test API
This adds a test API to check whether the IME Switcher button should be
shown when the IME is shown. This will be used in the tests for the
onCustomImeSwitcherButtonRequestedVisible API, added in [1], to check
whether the button is visible at the start of the test.

This check is done as an assumption rather than an assertion, to relax
the restriction against OEM changes in computing the IME Switcher button
requested visibility.

  [1]: I87e87cd30e45f3eb73b009a13cffb1676e067590

Flag: EXEMPT test API
Bug: 384484869
Test: atest InputMethodServiceTest#testOnCustomImeSwitcherButtonRequestedVisible_gestureNav
  InputMethodServiceTest#testOnCustomImeSwitcherButtonRequestedVisible_threeButtonNav
Change-Id: I168213cd24abec1ebe4c9943008b91f3d6122aef