Enable secondary_user_on_secondary_display for CtsShortcutManagerTestCases
secondary_user_on_secondary_display is for background users that have
access to UI on assigned displays (a.k.a. visible background users) on
devices that have config_multiuserVisibleBackgroundUsers enabled.
The main use case is Automotive's multi-display Whole Cabin experience
where passengers (modeled as visible background users) can interact
with the display in front of them concurrently with the driver
(modeled as the the current user) interacting with driver's display.

Fixes: 362860357
Test: atest CtsShortcutManagerTestCases
Test: atest --user-type secondary_user CtsShortcutManagerTestCases
Test: atest --user-type secondary_user_on_secondary_display CtsShortcutManagerTestCases
Flag: TEST_ONLY

Change-Id: Iec021784f790c4597cbf88d5d4ee791684af041c