AICS: API coverage
This will tweak the tool to believe there is actual coverage.
This is using a mock interface of the AudioInputControl as this require
having a remote device that would act as an HearingAid while running cts
test.
This is not possible, for both virtual device and physical device.

The behavior test of the AOSP implementation is guarantee with the
pts-bot. OEM will have to validate the stack behavior with the pts test

Fix: 385014991
Fix: 385014725
Fix: 385015367
Fix: 385014767
Fix: 385015159
Fix: 385015441
Fix: 385015323
Fix: 385015418
Fix: 385015366
Fix: 385014716
Fix: 385014955
Fix: 385015128
Fix: 385015021
Fix: 385015063
Fix: 385015417
Fix: 385014954
Fix: 385015365
Fix: 385015440
Fix: 385015155
Fix: 385014795
Fix: 385014715
Fix: 385015265
Fix: 385014885
Fix: 385015380
Test: atest CtsBluetoothTestCases
Change-Id: I6f41aef0be3f35d93a3e5230be50687506e9fcf5