Autofill No need to wait for window changed with requestFocus
In testCancelMultipleTimes(),
change focus to password for prepare for next step.
If requestFocusOnPassword() is used, wait for window changed.
But simple focus movement does not cause wondow changed.
Waiting for window changed here is not appropriate
and this is not necessary for the test.
Change it to only request focus to the password editbox.

Bug: 384599018
Test: run cts -m CtsAutoFillServiceTestCases -t
android.autofillservice.cts.dropdown.
LoginActivityTest#testCancelMultipleTimes
Change-Id: I9d6b7282caaac0ea07fe6a1cb26f447e82b6c340

Change-Id: Ie5c081295047e261a560660a601fb7ccaf565a77