Fix android.security.cts.MotionEventTest for secondary_user_on_secondary_display
secondary_user_on_secondary_display is for background users that have access to UI on assigned displays (a.k.a. visible background users) on devices that have config_multiuserVisibleBackgroundUsers enabled. The main use case is Automotive's multi-display Whole Cabin experience where passengers (modeled as visible background users) can interact with the display in front of them concurrently with the driver (modeled as the the current user) interacting with driver's display.

The fix is to specify the id of the display assigned to the user in the
MotionEvent.

Bug: 385358597
Flag: EXEMPT test fix only
Test: atest --user-type secondary_user_on_secondary_display android.security.cts.MotionEventTest
Test: atest android.security.cts.MotionEventTest
Change-Id: Ie2a599f164d37b3cb31165b6d5137ca01d8d3269