Added testOpenMixedMode to camera2 multi-client CTS
This test checks that when a java client opens the camera in shared mode
and a native client attempts to open it in exclusive mode, the native
client fails to open the camera. Likewise, when the native client opens
the camera in shared mode and the java client opens it in exclusive
mode, the native client is evicted from camera access.

Bug: 377371012
Test: atest SharedCameraTest
Change-Id: I8154729ba0e4070fdc7c17652863205a1fa7af33