Fix test flakiness in testDismissKeyguard.
The test KeyguardLockedTests#testDismissKeyguard was failing with a
8.51% failure rate because it was asserting the visibility of the
dismissing activity based on the WM state that keyguard was going away.
This commit adds a computeState call to force the computation of the
current window state. This ensures that the visibility assertion will
pass after the activity's visibility is properly updated.

Fix: 389216814
Test: atest CtsWindowManagerDeviceKeyguard:KeyguardLockedTests#testDismissKeyguard
Flag: TEST_ONLY
Change-Id: I3f4ab714bd7314595ebda983e04323329fc4b2fb