ITS: Util class for gen2 rig controllers.
This change includes the utitily functions to control the gen2 rig
servo and lighting controllers. Before the test starts, lights in the
rig will be turned ON and servo will be moved to center position.

Tested to see that lights turn ON and OFF.
Tested to see that the servo moves to desired position in both
clockwise and anti clockwise direction.

Bug: 380892830

Test: Tested with Pixel device in gen2 rig

FLAG: TEST_ONLY

Change-Id: I90273891d7e1595bb4929c69888b94e0229b706d