Fix throwing error for polling
Polling timeout does not make the test terminate with an error when the desired value was not achieved after set time. This leads to irrelevant errors down the line, where the main issue probably still is just wrong AppOpsMode. This commit aims to confirm that, by failing the test right after polling.

Bug: b/333230523
Flag: TEST_ONLY
Test: atest CtsDevicePolicyTestCases:android.devicepolicy.cts.CredentialManagementAppTest
Change-Id: I202fcba8eb964e3d74d6d23d609f22d26ab93c20