WebView: use messages in asserts inside for-loops
No change to logic. This only adds assertion messages to assert
statements that already existed. When an assertion fails in a for-loop,
it's difficult to know which iterate of the loop we are on. This adds
information about the loop iteration in the assertion message to make
debugging easier.

Bug: 372644502
Test: atest --rerun-until-failure=100 CtsWebkitTestCases:WebViewTest#testCanInjectHeaders
Test: atest --rerun-until-failure=100 CtsWebkitTestCases:WebViewTest#testGoBackAndForward
Flag: TEST_ONLY
Change-Id: I304f0caedc1ccb3abd8a4a3d286dff908ed9337e